+++ docs/Whats-New.md
@@ -208,6 +208,7 @@ Vanilla fixes:
 - Fix a bug where the AI would sell off buildings with `Artillary=yes`, `TickTank=yes` or `IsJuggernaut=yes` that had `UndeploysInto=none` when they were fired at by something outside of their weapon range (by Rampastring)
 - Fix a bug where harvesters on large maps could prefer unloading at refineries that were the longest distance away from the harvesters (by Rampastring)
 - Fix a bug where the camera kept following a followed object when a trigger or script told it to center on a waypoint or team (by Rampastring)
+- Fix a bug where crew wouldn't exit from construction yards when they were sold or destroyed (by ZivDero)
 
 </details>